requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
six==1.17.0
tesseract==0.1.3
urllib3==2.3.0